data/loader/requirements.txt
psycopg2-binary==2.8.6
pypika==0.48.1
